requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.4
langchain-ollama>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
gitpython>=3.1.40
python-dotenv>=1.0.0
